Properly save HTTP code on exceptions
Even if the code was passed as a keyword arg, it would show up as a 500
in the client because the default 'code' property is never updated from
its default value. It should be OK to pop this off kwargs; as it turns
out, no callers are actually using the message formatting functionality
provided here. Also no reason to store kwargs...