Don't fetch entire list when looking up by ID
Blazar APIs support directly fetching a resource (whether host, floating
ip, or lease) with its ID. However, the default logic for this includes
a block that supports additionally looking up the resource by name (to
allow looking up e.g., a lease by it's name.) This requires fetching the
entire list of all resources and iterating over them until a match is
found, which is very inefficient. Moreover, this branch would be taken
even if the input was already a UUID.

This commit changes that behavior so that if a UUID is provided, it is
used to directly fetch the resource, which seems to be what the original
intent was in the first place.

The 'name_key' on the Lease resource was also removed, as the value of
'name' is already a default.

As networks do not have a 'name' stored, they are now using
allow_names=False, similar to floating IPs.