Fix invalid column error message (#17)
This bug occurred when a user passed only invalid column names to the client via a list-* command. Column names are intended to be processed by the `cliff` library, which compares the base columns which may be viewed by the user against any columns they may have included in their arguments. There is an unnecessary processing step in the `setup_columns` method which removes any invalid columns input by the user from the base list. That means that if only invalid columns are included, `cliff` receives an empty list of possible columns, which it treats as the only invalid configuration. Otherwise, it will simply ignore any invalid columns in the args provided by the user.

This patch fixes the pre-processing step in `setup_columns` by ensuring that the base columns visible via the command will always be passed to cliff. The reason why this step was not eliminated entirely was to retain the previous behavior of the Blazar client where a user may "look ahead" at columns added to the Blazar API which are not yet supported by their client by simply passing them to the command with `-c`.