Add support for setting capability uniqueness in blazar client
Updated the `UpdateCapabilityCommand` in the blazar client to include a new `--unique` flag,
allowing users to set a capability as unique when updating or creating a capability.
Also, modified the `set_capability` method in the `ComputeHostClientManager` to include the
`is_unique` parameter when sending a PATCH request to update the capability.

This enhancement enables users to specify whether a capability should be unique or not via the blazar client.