Fix 'detail' parameter usage in Blazar client show command
This commit fixes the handling of the 'detail' parameter in the Blazar client show command. Since other resources like hosts and FIPs do not have a detail keyword in their manager get method, the previous implementation caused errors.

This commit removes the unnecessary usage of 'detail' and instead calls the 'additional_details' method if the parsed args have 'detail' arg to fetch additional details

Additionally, tests are added to cover the corrected behavior of the show command with and without the 'detail' parameter.