Fix DeviceShow and NetworkShow commands
Since commit 6c0ae07ee2426115fb080a32e2f8409e75d46375, the show command
parser moved away from the base class, and so we need to implement it in
our custom resource types Network and Device. This also casts the
field we are comparing to string so that segment id lookup will work.